add webpack bundler
