+++ package.json
@@ -1,7 +1,6 @@
 {
-  "scripts":{
-    "build": "tsc --build",
-    "clean": "tsc --build --clean"
+  "scripts": {
+    "build": "webpack --mode production"
   },
   "dependencies": {
     "graphql": "^16.6.0",
@@ -9,6 +8,10 @@
   },
   "devDependencies": {
     "@types/chrome": "^0.0.278",
-    "typescript": "^5.6.3"
+    "copy-webpack-plugin": "^12.0.2",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.6.3",
+    "webpack": "^5.95.0",
+    "webpack-cli": "^5.1.4"
   }
 }
