+++ src/manifest.json
@@ -2,7 +2,7 @@
   "manifest_version": 3,
   "name": "uOttawaRMP",
   "description": "RMP extension for uOttawa's course page.",
-  "version": "1.1.0",
+  "version": "0.1.0",
   "action": {
     "default_title": "Search RateMyProf"
   },
