+++ tsconfig.json
@@ -9,6 +9,7 @@
     "lib": ["es2015", "dom"]
   },
   "include": [
-    "src/**/*"
+    "src/**/*",
+    "manifest.json"
   ]
 }
